Adaptive number of interpolation points
In trk_compile_data.m, you can now automatically choose the number of interpolation points based on the tract lengths of the first subject. This is determined by a new 'nPts' column in tract_info.txt that takes the place of the old nPts input argument.

If nPts is empty ([]) for a given tract, trk_compile_data.m will determine the # of interpolation points based on the mean streamline length of the first subject. This number will then be used for the rest of the subjects.

Since interpolated tracks can now have differing numbers of vertices, track_means and scalar_means are now combined in a structure form instead of being matrices.

Exchanged the for loops over subjects and tracts to make this easier. Also allowed the QC fig printing to be brought inside the loops.